Bug #404284 - add autotest module int sprdroid5.0_trunk platform
[bug number  ] 404284
[root cause  ] autotest module int sprdroid5.0_trunk platform dosen't work
[changes     ] add autotest module int sprdroid5.0_trunk platform
[side effects] none
[self test   ] board: 8730 mobiletester/writeimei/autotester, log test ok
[whether AOB ]
[reviewers   ] alvin.zhou

+++ autotest/camera.cpp
@@ -7,7 +7,7 @@
 #include <errno.h>
 #include <fcntl.h>
 #include <pthread.h>
-#include <asm/page.h>
+//#include <asm/page.h>
 #include <linux/ion.h>
 #include <linux/videodev2.h>
 #include <sys/ioctl.h>
@@ -432,7 +432,7 @@ static int dcam_alloc_mem( void )
 		return -errno;
 	}
 
-	int phy_addr = ipdata.phys;
+	unsigned long phy_addr = ipdata.phys;
 	int phy_size = ipdata.size;
 
 	uchar * viraddr = (uchar *)mmap(0, mem_size, PROT_READ | PROT_WRITE,
